requirements.txt
flask
gunicorn
yfinance
apscheduler
python-dotenv
requests
pyTelegramBotAPI
pandas
